requirements.txt
pandas>=1.3.0,<2.0.0
numpy>=1.20.0,<1.25.0
ccxt>=4.0.0,<4.1.0
tenacity>=8.0.0
python-dotenv>=0.19.0
requests>=2.25.0
websocket-client>=1.2.0
flask>=2.0.0,<3.0.0
TA-Lib>=0.4.28
plotly>=5.0.0
gunicorn>=20.0.0